fhe -> seal -> afhe * UML Diagram
